Move test files to tst/model directory
