+++ tst/model/HexCalculatorTest.java
@@ -1,8 +1,7 @@
-package utils;
+package model;
 
-import model.HexCalculator;
-import model.Hexadecimal;
 import org.junit.jupiter.api.Test;
+
 import static org.junit.jupiter.api.Assertions.assertEquals;
 class HexCalculatorTest {
